rename: HomeCategory 폴더의 파일을 components 폴더로 이동
