+++ src/components/CategoryCard.tsx
